feat: criacao cargo ajuste de todos os services

+++ app/empresas/api.py
@@ -2,46 +2,57 @@ from ninja import Router
 from app.empresas.service import EmpresasService
 from app.empresas.schemas import EmpresaSchemaOut, EmpresaSchemaIn, EmpresaSchemaAuto
 from app.authenticate.service import JWTAuth
+from app.utils.jwt_manager import authenticate
+from django.http import JsonResponse
 
-router = Router(auth=JWTAuth(), tags=['Empresas'])
+
+empresas_router = Router(auth=JWTAuth(), tags=['Empresas'])
 service = EmpresasService ()
 
 #GETS
-@router.get("", response=list[EmpresaSchemaOut])
+@empresas_router.get("", response=list[EmpresaSchemaOut], auth=None)
 def get_empresa(request):
-    response = service.get_empresa()    
-    return response
+    return service.get_empresa()    
+    
 
-@router.get("{empresa_id}", response=EmpresaSchemaOut)
+@empresas_router.get("{empresa_id}", response=EmpresaSchemaOut, auth=None)
 def get_empresas_by_id(request, empresa_id: str):
-    response = service.get_empresas_by_id(id = empresa_id)
-    return response
+    return service.get_empresas_by_id(id = empresa_id)
+    
 
-@router.get("autocomplete/{cnpj}", response=EmpresaSchemaAuto)
+@empresas_router.get("autocomplete/{cnpj}", response=EmpresaSchemaAuto, auth=None)
 def autocomplete_empresa(request, cnpj: str):
-    response = service.autocomplete_empresa(request, cnpj=cnpj)
-    return response
+    return service.autocomplete_empresa(request, cnpj=cnpj)
+    
 
 #POST
-@router.post("", auth=None)
+@empresas_router.post("")
 def create_empresa(request, payload: EmpresaSchemaIn):
-    response = service.create_empresa(payload)
-    return response
-
+    return service.create_empresa(request, payload)
+    
 
 #PATCH
-@router.patch("{empresa_id}")
+@empresas_router.patch("{empresa_id}")
 def update_empresa(request, empresa_id: str, payload: EmpresaSchemaIn):
-    response = service.update_empresa(request, empresa_id, payload)
-    return response
+    token = authenticate(request)    
+    if not token.get("empresa_id") == empresa_id and token.get("is_superuser") == True:
+        return JsonResponse(data={'error': "usuário inválido"}, status=400)
+    return service.update_empresa(request, empresa_id, payload)
+    
 
 #DELETE
-@router.delete("delete/{empresa_id}")
+@empresas_router.delete("delete/{empresa_id}")
 def soft_delete_empresa(request, empresa_id: str):
-    response = service.soft_delete_empresa(empresa_id)
-    return response
+    token = authenticate(request)    
+    if not token.get("empresa_id") == empresa_id and token.get("is_superuser") == True:
+        return JsonResponse(data={'error': "usuário inválido"}, status=400)
+    return service.soft_delete_empresa(empresa_id)
+    
 
-@router.delete("{empresa_id}")
+@empresas_router.delete("{empresa_id}")
 def delete_empresa(request, empresa_id: str):
-    response = service.delete_empresa(empresa_id)
-    return response
+    token = authenticate(request)    
+    if not token.get("empresa_id") == empresa_id and token.get("is_superuser") == True:
+        return JsonResponse(data={'error': "usuário inválido"}, status=400)
+    return service.delete_empresa(empresa_id)
+
